Struct soa_derive_example::ParticleVec

pub struct ParticleVec {
    pub mass: Vec<f64>,
    pub position: Vec<[f64; 3]>,
    pub kind: Vec<usize>,
    pub name: Vec<String>,
}

An analog to Vec<Particle> with Struct of Array (SoA) layout

Fields
mass: Vec<f64>

A vector of mass from a Particle

position: Vec<[f64; 3]>

A vector of position from a Particle

kind: Vec<usize>

A vector of kind from a Particle

name: Vec<String>

A vector of name from a Particle

Methods

impl ParticleVec

fn iter(&self) -> Iter

Get an iterator over the ParticleRef in this vector

impl ParticleVec

fn iter_mut(&mut self) -> IterMut

Get a mutable iterator over the ParticleRefMut in this vector

impl ParticleVec

fn zip<'a, D>(&'a self, data: D) -> Self::Iterator where
Self: Zip<'a, D>,

Get an iterator over multiple fields of this ParticleVec.

Use the data argument to request specific fields. This argument can be a reference to any of the marker types in the zip_particle module; or a tuple of reference to types in this module. Only fields marked with #[soa_derive(zip)] are available. It is also possible to pass any type implementing IntoIterator after the marker types, and the corresponding values will be forwarded.

This function is only generated when the main struct is public.

Examples

 #[derive(StructOfArray)]
 pub struct Example {
     #[soa_derive(zip)]
     mass: f64,
     #[soa_derive(zip)]
     label: String,
     #[soa_derive(zip)]
     position: [f64; 3],
 }

 use self::zip_example::{Mass, Label, Position};
 let vector = ExampleVec::new();
 for mass in vector.zip(&Mass) {
     // ...
 }

 for (mass, position, label) in vector.zip((&Mass, &Position, &Label)) {
     // ...
 }

 // Using other data in the `zip` function
 let data = vec![3.3; 4];
 for (mass, data) in vector.zip((&Mass, &data)) {
     // ...
 }

fn zip_mut<'a, D>(&'a mut self, data: D) -> Self::Iterator where
Self: ZipMut<'a, D>,

Get an iterator over multiple fields of this ParticleVec.

Use the data argument to request specific fields. This argument can be a reference or a mutable reference to any of the marker types in the zip_particle module; or a tuple of reference or mutable reference to types in this module. Only fields marked with #[soa_derive(zip)] are available. It is also possible to pass any type implementing IntoIterator after the marker types, and the corresponding values will be forwarded.

This function is only generated when the main struct is public.

Examples

 #[derive(StructOfArray)]
 pub struct Example {
     #[soa_derive(zip)]
     mass: f64,
     #[soa_derive(zip)]
     label: String,
     #[soa_derive(zip)]
     position: [f64; 3],
 }

 use self::zip_example::{Mass, Label, Position};
 let mut vector = ExampleVec::new();
 for mass in vector.zip_mut(&mut Mass) {
     *mass += 3.0;
 }

 // Mutable position, immutable mass and label
 for (&mass, position, label) in vector.zip_mut((&Mass, &mut Position, &Label)) {
     if label == "foo" {
         position[0] -= mass;
     }
 }

 // Using other data in the `zip` function
 let data = vec![3.3; 4];
 for (mass, data) in vector.zip((&Mass, &data)) {
     // ...
 }

impl ParticleVec

fn new() -> ParticleVec

Similar to Vec<Particle> ::new()

fn with_capacity(capacity: usize) -> ParticleVec

Similar to Vec<Particle> ::with_capacity(), initializing all fields with the given capacity.

fn capacity(&self) -> usize

Similar to Vec<Particle> ::capacity(), the capacity of all fields should be the same.

fn reserve(&mut self, additional: usize)

Similar to Vec<Particle> ::reserve(), reserving the same additional space for all fields.

fn reserve_exact(&mut self, additional: usize)

Similar to Vec<Particle> ::reserve_exact() reserving the same additional space for all fields.

fn shrink_to_fit(&mut self)

Similar to Vec<Particle> ::shrink_to_fit() shrinking all fields.

fn truncate(&mut self, len: usize)

Similar to Vec<Particle> ::truncate() truncating all fields.

fn push(&mut self, value: Particle)

Similar to Vec<Particle> ::push().

fn len(&self) -> usize

Similar to Vec<Particle> ::len(), all the fields should have the same length.

fn is_empty(&self) -> bool

Similar to Vec<Particle> ::is_empty(), all the fields should have the same length.

fn swap_remove(&mut self, index: usize) -> Particle

Similar to Vec<Particle> ::swap_remove().

fn insert(&mut self, index: usize, element: Particle)

Similar to Vec<Particle> ::insert().

fn remove(&mut self, index: usize) -> Particle

Similar to Vec<Particle> ::remove().

fn pop(&mut self) -> Option<Particle>

Similar to Vec<Particle> ::pop().

fn append(&mut self, other: &mut ParticleVec)

Similar to Vec<Particle> ::append().

fn clear(&mut self)

Similar to Vec<Particle> ::clear().

fn split_off(&mut self, at: usize) -> ParticleVec

Similar to Vec<Particle> ::split_off().

fn as_slice(&self) -> ParticleSlice

Similar to Vec<Particle> ::as_slice().

fn as_mut_slice(&mut self) -> ParticleSliceMut

Similar to Vec<Particle> ::as_mut_slice().

fn slice(&self, range: Range<usize>) -> ParticleSlice

Create a slice of this vector matching the given range. This is analogous to Index<Range<usize>>.

fn slice_mut(&mut self, range: Range<usize>) -> ParticleSliceMut

Create a mutable slice of this vector matching the given range. This is analogous to IndexMut<Range<usize>>.

fn retain<F>(&mut self, f: F) where
F: FnMut(ParticleRef) -> bool,

Similar to Vec<Particle> ::retain().