Module soa_derive_example::zip_particle

This module contains the marker types for the ParticleVec::zip() functionality.

Structs

Kind
Marker type to access the kind field of a ParticleVec in the ParticleVec::zip() function and familly

Mass
Marker type to access the mass field of a ParticleVec in the ParticleVec::zip() function and familly

Position
Marker type to access the position field of a ParticleVec in the ParticleVec::zip() function and familly