soa_derive_example

Struct ParticleSlice

pub struct ParticleSlice<'a> {
    pub mass: &'a [f64],
    pub position: &'a [[f64; 3]],
    pub kind: &'a [usize],
    pub name: &'a [String],
}

A slice of Particle inside a ParticleVec .

Fields
mass: &'a [f64]

slice of mass inside a ParticleVec

position: &'a [[f64; 3]]

slice of position inside a ParticleVec

kind: &'a [usize]

slice of kind inside a ParticleVec

name: &'a [String]

slice of name inside a ParticleVec

Implementations

impl<'a> ParticleSlice<'a>

pub fn len(&self) -> usize

Similar to & [Particle] ::len(), the length of all fields should be the same.

pub fn is_empty(&self) -> bool

Similar to & [Particle] ::is_empty(), the length of all fields should be the same.

pub fn first(&self) -> Option<ParticleRef<'a>>

Similar to & [Particle] ::first().

pub fn split_first(&self) -> Option<(ParticleRef<'a>, ParticleSlice<'a>)>

Similar to & [Particle] ::split_first().

pub fn last(&self) -> Option<ParticleRef<'a>>

Similar to & [Particle] ::last().

pub fn split_last(&self) -> Option<(ParticleRef<'a>, ParticleSlice<'a>)>

Similar to & [Particle] ::split_last().

pub fn split_at(&self, mid: usize) -> (ParticleSlice<'a>, ParticleSlice<'a>)

Similar to & [Particle] ::split_at().

pub fn get<'b, I>(&'b self, index: I) -> Option<I::RefOutput>
where I: SoAIndex<ParticleSlice<'b>>, 'a: 'b,

Similar to & [Particle] ::get().

pub unsafe fn get_unchecked<'b, I>(&'b self, index: I) -> I::RefOutput
where I: SoAIndex<ParticleSlice<'b>>, 'a: 'b,

Similar to & [Particle] ::get_unchecked().

pub fn index<'b, I>(&'b self, index: I) -> I::RefOutput
where I: SoAIndex<ParticleSlice<'b>>, 'a: 'b,

Similar to the std::ops::Index trait for & [Particle] . This is required because we cannot implement std::ops::Index directly since it requires returning a reference.

pub fn reborrow<'b>(&'b self) -> ParticleSlice<'b>
where 'a: 'b,

Reborrows the slices in a narrower lifetime

pub fn as_ptr(&self) -> ParticlePtr

Similar to & [Particle] ::as_ptr().

pub unsafe fn from_raw_parts<'b>( data: ParticlePtr, len: usize, ) -> ParticleSlice<'b>

Similar to std::slice::from_raw_parts().

impl<'a> ParticleSlice<'a>

pub fn iter(&self) -> ParticleIter<'_>

Get an iterator over the ParticleRef in this slice.

pub fn into_iter(self) -> ParticleIter<'a>

Get an iterator over the ParticleRef in this slice.