Struct soa_derive_example::ParticleSlice

pub struct ParticleSlice<'a> {
    pub mass: &'a [f64],
    pub position: &'a [[f64; 3]],
    pub kind: &'a [usize],
    pub name: &'a [String],
}

A slice of Particle inside a ParticleVec .

Fields
mass: &'a [f64]

A slice of mass from a ParticleVec

position: &'a [[f64; 3]]

A slice of position from a ParticleVec

kind: &'a [usize]

A slice of kind from a ParticleVec

name: &'a [String]

A slice of name from a ParticleVec

Methods

impl<'a> ParticleSlice<'a>

fn iter(&self) -> Iter

Get an iterator over the ParticleRef in this slice.

impl<'b> ParticleSlice<'b>

fn zip<'a, D>(&'a self, data: D) -> Self::Iterator where
Self: Zip<'a, D>,

Get an iterator over multiple fields of this ParticleSlice.

Use the data argument to request specific fields. This argument can be a reference to any of the marker types in the zip_particle module; or a tuple of reference to types in this module. Only fields marked with #[soa_derive(zip)] are available. It is also possible to pass any type implementing IntoIterator after the marker types, and the corresponding values will be forwarded.

This function is only generated when the main struct is public.

Examples

 #[derive(StructOfArray)]
 pub struct Example {
     #[soa_derive(zip)]
     mass: f64,
     #[soa_derive(zip)]
     label: String,
     #[soa_derive(zip)]
     position: [f64; 3],
 }

 use self::zip_example::{Mass, Label, Position};
 let vector = ExampleVec::new();

 let slice = vector.as_slice();
 for mass in slice.zip(&Mass) {
     // ...
 }

 for (mass, position, label) in slice.zip((&Mass, &Position, &Label)) {
     // ...
 }

impl<'a> ParticleSlice<'a>

fn len(&self) -> usize

Similar to [Particle] ::len(), the length of all fields should be the same.

fn is_empty(&self) -> bool

Similar to [Particle] ::is_empty(), the length of all fields should be the same.

fn first(&self) -> Option<ParticleRef>

Similar to [Particle] ::first().

fn split_first(&self) -> Option<(ParticleRef, ParticleSlice)>

Similar to [Particle] ::split_first().

fn last(&self) -> Option<ParticleRef>

Similar to [Particle] ::last().

fn split_last(&self) -> Option<(ParticleRef, ParticleSlice)>

Similar to [Particle] ::split_last().

fn split_at(&self, mid: usize) -> (ParticleSlice, ParticleSlice)

Similar to [Particle] ::split_at().

fn get(&self, i: usize) -> Option<ParticleRef>

Similar to [Particle] ::get().

unsafe fn get_unchecked(&self, i: usize) -> ParticleRef

Similar to [Particle] ::get_unchecked().