Struct soa_derive_example::ParticleSliceMut

pub struct ParticleSliceMut<'a> {
    pub mass: &'a mut [f64],
    pub position: &'a mut [[f64; 3]],
    pub kind: &'a mut [usize],
    pub name: &'a mut [String],
}

A mutable slice of Particle inside a ParticleVec .

Fields
mass: &'a mut [f64]

A mutable slice of mass from a ParticleVec

position: &'a mut [[f64; 3]]

A mutable slice of position from a ParticleVec

kind: &'a mut [usize]

A mutable slice of kind from a ParticleVec

name: &'a mut [String]

A mutable slice of name from a ParticleVec

Implementations

impl<'a> ParticleSliceMut<'a>

pub fn iter(&mut self) -> Iter<'_>

Get an iterator over the ParticleRef in this vector

pub fn iter_mut(&mut self) -> IterMut<'_>

Get a mutable iterator over the ParticleRefMut in this vector

impl<'a> ParticleSliceMut<'a>

pub fn as_ref(&self) -> ParticleSlice<'_>

Convert a ParticleSliceMut to a ParticleSlice in order to be able to use the methods on the non mutable version of the slices.

pub fn len(&self) -> usize

Similar to [Particle] ::len(), the length of all fields should be the same.

pub fn is_empty(&self) -> bool

Similar to [Particle] ::is_empty(), the length of all fields should be the same.

pub fn first_mut(&mut self) -> Option<ParticleRefMut<'_>>

Similar to [Particle] ::first_mut().

pub fn split_first_mut(
&mut self
) -> Option<(ParticleRefMut<'_>, ParticleSliceMut<'_>)>

Similar to [Particle] ::split_first_mut().

pub fn last_mut(&mut self) -> Option<ParticleRefMut<'_>>

Similar to [Particle] ::last_mut().

pub fn split_last_mut(
&mut self
) -> Option<(ParticleRefMut<'_>, ParticleSliceMut<'_>)>

Similar to [Particle] ::last_mut().

pub fn split_at_mut(
&mut self,
mid: usize
) -> (ParticleSliceMut<'_>, ParticleSliceMut<'_>)

Similar to [Particle] ::split_at_mut().

pub fn swap(&mut self, a: usize, b: usize)

Similar to [Particle] ::swap().

pub fn get<'b, I>(&'b self, index: I) -> Option<I::RefOutput> where
I: SoAIndex<ParticleSlice<'b>>,
'a: 'b,

Similar to [Particle] ::get().

pub unsafe fn get_unchecked<'b, I>(&'b self, index: I) -> I::RefOutput where
I: SoAIndex<ParticleSlice<'b>>,
'a: 'b,

Similar to [Particle] ::get_unchecked().

pub fn index<'b, I>(&'b self, index: I) -> I::RefOutput where
I: SoAIndex<ParticleSlice<'b>>,
'a: 'b,

Similar to std::ops::Index trait on [Particle] . This is required because we cannot implement that trait.

pub fn get_mut<'b, I>(&'b mut self, index: I) -> Option<I::MutOutput> where
I: SoAIndexMut<ParticleSliceMut<'b>>,
'a: 'b,

Similar to [Particle] ::get_mut().

pub unsafe fn get_unchecked_mut<'b, I>(&'b mut self, index: I) -> I::MutOutput where
I: SoAIndexMut<ParticleSliceMut<'b>>,
'a: 'b,

Similar to [Particle] ::get_unchecked_mut().

pub fn index_mut<'b, I>(&'b mut self, index: I) -> I::MutOutput where
I: SoAIndexMut<ParticleSliceMut<'b>>,
'a: 'b,

Similar to std::ops::IndexMut trait on [Particle] . This is required because we cannot implement std::ops::IndexMut directly since it requires returning a mutable reference.

pub fn as_slice<'b>(&'b self) -> ParticleSlice<'b> where
'a: 'b,

Returns a non-mutable slice from this mutable slice.

pub fn reborrow<'b>(&'b mut self) -> ParticleSliceMut<'b> where
'a: 'b,

Reborrows the slices in a narrower lifetime

pub fn as_ptr(&self) -> ParticlePtr

Similar to [Particle] ::as_ptr().

pub fn as_mut_ptr(&mut self) -> ParticlePtrMut

Similar to [Particle] ::as_mut_ptr().

pub unsafe fn from_raw_parts_mut<'b>(
data: ParticlePtrMut,
len: usize
) -> ParticleSliceMut<'b>

Similar to std::slice::from_raw_parts_mut().