Struct soa_derive_example::ParticleSliceMut

pub struct ParticleSliceMut<'a> {
    pub mass: &'a mut [f64],
    pub position: &'a mut [[f64; 3]],
    pub kind: &'a mut [usize],
    pub name: &'a mut [String],
}

A mutable slice of Particle inside a ParticleVec .

Fields
mass: &'a mut [f64]

A mutable slice of mass from a ParticleVec

position: &'a mut [[f64; 3]]

A mutable slice of position from a ParticleVec

kind: &'a mut [usize]

A mutable slice of kind from a ParticleVec

name: &'a mut [String]

A mutable slice of name from a ParticleVec

Methods

impl<'a> ParticleSliceMut<'a>

fn iter(&mut self) -> Iter

Get an iterator over the ParticleRef in this vector

fn iter_mut(&mut self) -> IterMut

Get a mutable iterator over the ParticleRefMut in this vector

impl<'b> ParticleSliceMut<'b>

fn zip<'a, D>(&'a self, data: D) -> Self::Iterator where
Self: Zip<'a, D>,

Get an iterator over multiple fields of this ParticleSliceMut.

Use the data argument to request specific fields. This argument can be a reference to any of the marker types in the zip_particle module; or a tuple of reference to types in this module. Only fields marked with #[soa_derive(zip)] are available. It is also possible to pass any type implementing IntoIterator after the marker types, and the corresponding values will be forwarded.

This function is only generated when the main struct is public.

Examples

 #[derive(StructOfArray)]
 pub struct Example {
     #[soa_derive(zip)]
     mass: f64,
     #[soa_derive(zip)]
     label: String,
     #[soa_derive(zip)]
     position: [f64; 3],
 }

 use self::zip_example::{Mass, Label, Position};
 let mut vector = ExampleVec::new();

 let slice = vector.as_mut_slice();
 for mass in slice.zip(&Mass) {
     // ...
 }

 for (mass, position, label) in slice.zip((&Mass, &Position, &Label)) {
     // ...
 }

fn zip_mut<'a, D>(&'a mut self, data: D) -> Self::Iterator where
Self: ZipMut<'a, D>,

Get an iterator over multiple fields of this ParticleSliceMut.

Use the data argument to request specific fields. This argument can be a reference or a mutable reference to any of the marker types in the zip_particle module; or a tuple of reference or mutable reference to types in this module. Only fields marked with #[soa_derive(zip)] are available. It is also possible to pass any type implementing IntoIterator after the marker types, and the corresponding values will be forwarded.

This function is only generated when the main struct is public.

Examples

 #[derive(StructOfArray)]
 pub struct Example {
     #[soa_derive(zip)]
     mass: f64,
     #[soa_derive(zip)]
     label: String,
     #[soa_derive(zip)]
     position: [f64; 3],
 }

 use self::zip_example::{Mass, Label, Position};
 let mut vector = ExampleVec::new();

 let mut slice = vector.as_mut_slice();
 for mass in slice.zip_mut(&mut Mass) {
     *mass += 3.0;
 }

 // Mutable position, immutable mass and label
 for (&mass, position, label) in slice.zip_mut((&Mass, &mut Position, &Label)) {
     if label == "foo" {
         position[0] -= mass;
     }
 }

impl<'a> ParticleSliceMut<'a>

fn as_ref(&self) -> ParticleSlice

Convert a ParticleSliceMut to a ParticleSlice in order to be able to use the methods on the non mutable version of the slices.

fn len(&self) -> usize

Similar to [Particle] ::len(), the length of all fields should be the same.

fn is_empty(&self) -> bool

Similar to [Particle] ::is_empty(), the length of all fields should be the same.

fn first_mut(&mut self) -> Option<ParticleRefMut>

Similar to [Particle] ::first_mut().

fn split_first_mut(&mut self) -> Option<(ParticleRefMut, ParticleSliceMut)>

Similar to [Particle] ::split_first_mut().

fn last_mut(&mut self) -> Option<ParticleRefMut>

Similar to [Particle] ::last_mut().

fn split_last_mut(&mut self) -> Option<(ParticleRefMut, ParticleSliceMut)>

Similar to [Particle] ::last_mut().

fn split_at_mut(&mut self, mid: usize) -> (ParticleSliceMut, ParticleSliceMut)

Similar to [Particle] ::split_at_mut().

fn swap(&mut self, a: usize, b: usize)

Similar to [Particle] ::swap().

fn get(&self, i: usize) -> Option<ParticleRef>

Similar to [Particle] ::get().

unsafe fn get_unchecked(&self, i: usize) -> ParticleRef

Similar to [Particle] ::get_unchecked().

fn get_mut(&mut self, i: usize) -> Option<ParticleRefMut>

Similar to [Particle] ::get_mut().

unsafe fn get_unchecked_mut(&mut self, i: usize) -> ParticleRefMut

Similar to [Particle] ::get_unchecked_mut().